Add function operate and declare variables
This commit adds the function operate, which takes three arguments to
operate on two numbers and call one of the four functions created in
the previous commit depending on which operator is passed into the
middle argument. It also declares three variables which can be passed
into the function later on.