Add functionality to operator buttons
This commit adds event listeners for the plus, minus, times, divide by,
equals, and clear buttons. Pressing an operator takes the current
display text content and assigns it to the first operand, then assigns
the appropriate operator to the operator variable. Pressing the equals
button assigns the display text content to the second operand, then
runs the operate function. Pressing the clear button clears all three
variables and sets the display text to 0. At the moment, there is a
problem where pressing an operator button will leave the first number
on the screen, so that is the next thing to fix.