Add full functionality to calculator
This commit gives full functionality to the calculator. It resets the
display after an operator or equals sign is pressed thanks to the flag
shouldResetDisplay, and uses callback functions handleOperator and
handleEquals to utilize better logic for operations. It also features a
rounding mechanic, a safeguard against consecutive operator presses,
and a snarky error message when dividing by zero. Barring additional
extra credit features, the calculator is complete.